Add mongomock to project, change init collections
In addition to adding the mongomock package to the project's
requirements.txt, this commit also alters the init_collections func
(previously init_schemas) to allow for specififying whether or not JSON
Schema validators will be included. This accounts for the current lack
of special options being implmented in v4.3.0 of mongomock.