Mongo & Create Application Testing (#11)
* Add mongomock to project, change init collections

In addition to adding the mongomock package to the project's
requirements.txt, this commit also alters the init_collections func
(previously init_schemas) to allow for specififying whether or not JSON
Schema validators will be included. This accounts for the current lack
of special options being implmented in v4.3.0 of mongomock.

* Add success & failure case tests for CreateApplication

* Add integration and markers

Adds integration markers through project pytest.ini file.
Adds unit tests for expected fail cases in the create_application
endpoint.
Adds functions for integration tests for persistence and schema
functionality on Mongo end.

* Add integration tests, deployed db injection

Commits adds both an integration test for persistence of the application
document following a POST API call and for rejection of a forced,
external-of-API insert of an application document. Also adds another
fixture implementation for the injection of the deployed database under
exclusively for testing.