Add pydantic model unit tests, fix model extra def
Changes the definition of PathwayGoals and Couse base models to not
allow for extra fields to better reflect design doc. Adds integration
tests for 'courses' collection schema acceptances and rejections. Adds
unit tests to account for expected success, failure, and results of
pydantic model validations for these new collection documents. These are
done to account for a lack existing endpoints.