Add Accelerate Flex, Pathway Goals, & Courses Collections (#19)
* Add fields to pydantic models & schema properties to collections

* Add pydantic model unit tests

Changes the definition of PathwayGoals and Couse base models to not
allow for extra fields to better reflect design doc. Adds integration
tests for 'courses' collection schema acceptances and rejections. Adds
unit tests to account for expected success, failure, and results of
pydantic model validations for these new collection documents. These are
done to account for a lack existing endpoints.